SCALE-SEAM (90) — the gate caught what my own check was built to miss
Eight methods out of client.ts (906 -> 866) into a new clientPortal.ts.

WHAT THEY ANSWER: how does someone with no account see this project, and
answer back? The owner's half mints, lists and revokes share tokens and reads
what came back; the recipient's half is the token-authenticated public
surface — the no-login page and digest URLs, and posting a comment or an
approve/acknowledge/decline. The TOKEN is the seam: minted on one side, IS
the credential on the other, and revoking it closes both. Splitting
owner-side from recipient-side would put a capability and its only means of
exercise in different files.

aec_api/routers/client_portal.py groups the same set — 8 of its 9 routes,
checked rather than assumed. The ninth, GET /shared/{token}/model.frag, has
no client method BY DESIGN: the server-rendered share page fetches it, not
this SPA.

THE EXTRACTION SCRIPT HAD A REAL BUG, AND A TEST FOUND IT RATHER THAN I DID.
Its brace counter only terminated when j > i, so a ONE-LINE method consumed
one line too many and swallowed the next method's doc comment. sharedPageUrl
and sharedDigestUrl are one-liners: the first draft dragged
spaceUtilBenchmarks's SPACE-UTIL doc out of client.ts and duplicated
sharedPageUrl's. docComments.test.ts went red, the extraction was reverted,
the counter fixed to terminate on the method line itself, and the slice
redone.

I HAD WRITTEN AN ORPHAN CHECK OF MY OWN AND IT PASSED. It treated a following
comment as an acceptable neighbour — precisely the case that defines the
defect. The real gate asserts that a line ending */ is not followed by a line
starting /**, and that is the whole difference. A check written while looking
at the code you just wrote inherits its blind spot.

Seven of these eight carry /** */ docs, which the walk-back did not handle
either until this slice; (88) and (89) survived only because the methods they
moved used // banners. Luck, not care. Both are fixed now.

FOUND ON THE WAY, DELIBERATELY NOT FIXED HERE: that ninth route serves
geometry only to a token minted with show_model, an opt-in the backend treats
as strictly independent of show_payments. createShareToken never sends it and
the row type shareTokens returns has no show_model field, so every token this
product mints has it false, /shared/{token}/model.frag always 404s, and the
public 3D viewer is dark from the UI's side. Left alone because an extraction
slice's claim is that no behaviour changed; recorded in the mixin header at
the method rather than only in a backlog.

101 above the banner now, still no map.

Verified: tsc exit 0, eslint exit 0, 203 files / 2,059 web tests, ruff clean
tree-wide, all seven doc and size gates OK.
